requirements.txt
flask==3.1.0
flask-cors==5.0.0
pillow==11.1.0
pytesseract==0.3.13
beautifulsoup4==4.12.3
requests==2.31.0
scikit-learn==1.4.1.post1
numpy==1.26.4
opencv-python==4.9.0.80
joblib==1.3.2
scipy==1.12.0 